Initial commit of this solution separated from CSF Software utils
This project is based on the old/original CSF Software Utilities project found
at https://github.com/CSF-Software-Development/csf-software-utilities

This commit breaks away the tracking of this subproject into a separate git
repository of its own.  It removes all data from this repository which does not
relate to the CSF.Data.NHibernate namespace and replaces the link to the CSF
Software utils project with a git submodule.

* Rename some of the solution files to make it more apparent that this solution
  is just the NHibernate-related functionality.