Resolve #38 - Fix issues, include file & add test coverage
This resolves the problem where a file was accidentally not included
in the compiled library.  It also had a few issues/bugs which were not
apparent previously (because it had not been built).

I have also added some happy-path test coverage for it as well, while
I was re-enabling it anyway.